refactor(scanner): give QrScannerModal its own file
The file already exported two components. Shared helpers move to
qr-scanner-shared.ts so neither component imports the other's module.
