+++ src/components/qr-scanner-panel.tsx
@@ -6,7 +6,7 @@ import {
   ScanLine,
   Trash2,
 } from "lucide-react"
-import { AppError, errorMessageKey, type ErrorCode } from "@/crypto/errors"
+import { AppError } from "@/crypto/errors"
 import type { MultipartScanSession } from "@/features/multipart-scan-session"
 import { formatFramePositions } from "@/features/presentation"
 import {
@@ -18,65 +18,22 @@ import {
   type QrScanHandle,
 } from "@/qr/decode"
 import type { TransferState } from "@/qr/multipart/transfer-state"
-import type { V2ArtifactType } from "@/schemas/domain"
+import {
+  acceptedPayloadLabel,
+  actualPayloadLabel,
+  deliveryError,
+  localized,
+  localizedErrorCode,
+  targetForPayload,
+  type LocalizedText,
+  type QrScannerPanelProps,
+  type ScannerTarget,
+} from "@/components/qr-scanner-shared"
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert"
 import { Button } from "@/components/ui/button"
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
-import {
-  Dialog,
-  DialogContent,
-  DialogTitle,
-  DialogTrigger,
-} from "@/components/ui/dialog"
 import { Progress } from "@/components/ui/progress"
-import { useI18n, type MessageKey, type Translate } from "@/i18n"
-import type { InterpolationValues } from "@/i18n/messages"
-import { cn } from "@/lib/utils"
-
-export type ScannerTarget = "message" | "symmetric-key" | "public-key"
-
-const TARGET_PREFIX: Record<ScannerTarget, string> = {
-  message: "OCM1:",
-  "symmetric-key": "OCK1:",
-  "public-key": "OCP1:",
-}
-
-const TARGET_LABEL_KEY: Record<ScannerTarget, MessageKey> = {
-  message: "scanner.targetLabel.message",
-  "symmetric-key": "scanner.targetLabel.symmetricKey",
-  "public-key": "scanner.targetLabel.publicKey",
-}
-
-export interface MultipartScanCompletion {
-  artifactType: V2ArtifactType
-  artifactBytes: Uint8Array
-}
-
-interface QrScannerPanelBaseProps {
-  singleTargets: ScannerTarget[]
-  onSingleScan: (
-    target: ScannerTarget,
-    payload: string,
-  ) => void | Promise<void>
-  cameraAvailable?: boolean
-  title?: string
-  autoStart?: boolean
-  stopHint?: string
-}
-
-type QrScannerPanelMultipartProps =
-  | {
-      multipart: {
-        session: MultipartScanSession
-        onComplete: (
-          completion: MultipartScanCompletion,
-        ) => void | Promise<void>
-      }
-    }
-  | { multipart?: never }
-
-export type QrScannerPanelProps = QrScannerPanelBaseProps &
-  QrScannerPanelMultipartProps
+import { useI18n } from "@/i18n"
 
 type ScannerMode = "idle" | "running" | "delivering" | "stopped"
 
@@ -92,53 +49,6 @@ interface ScannerRun {
 }
 
 const IDLE_TRANSFER_STATE: TransferState = { kind: "idle" }
-
-interface LocalizedText {
-  key: MessageKey
-  values?: InterpolationValues
-}
-
-function localized(
-  key: MessageKey,
-  values?: InterpolationValues,
-): LocalizedText {
-  return values === undefined ? { key } : { key, values }
-}
-
-function localizedErrorCode(code: ErrorCode): LocalizedText {
-  return localized(errorMessageKey(code))
-}
-
-function targetForPayload(payload: string): ScannerTarget | null {
-  for (const target of Object.keys(TARGET_PREFIX) as ScannerTarget[]) {
-    if (payload.startsWith(TARGET_PREFIX[target])) return target
-  }
-  return null
-}
-
-function actualPayloadLabel(payload: string, t: Translate): string {
-  const target = targetForPayload(payload)
-  return target === null
-    ? t("scanner.payloadLabel.foreign")
-    : t(TARGET_LABEL_KEY[target])
-}
-
-function acceptedPayloadLabel(
-  singleTargets: ScannerTarget[],
-  acceptsMultipart: boolean,
-  t: Translate,
-): string {
-  const labels = singleTargets.map((target) => t(TARGET_LABEL_KEY[target]))
-  if (acceptsMultipart) labels.push(t("scanner.acceptedLabel.multipart"))
-  return (
-    labels.join(t("scanner.acceptedLabel.separator")) ||
-    t("scanner.acceptedLabel.fallback")
-  )
-}
-
-function deliveryError(error: unknown): AppError {
-  return error instanceof AppError ? error : new AppError("INVALID_QR_PAYLOAD")
-}
 
 export function QrScannerPanel(props: QrScannerPanelProps) {
   const { language, t } = useI18n()
@@ -899,301 +809,5 @@ export function QrScannerPanel(props: QrScannerPanelProps) {
         )}
       </CardContent>
     </Card>
-  )
-}
-
-export type QrScannerModalProps = QrScannerPanelProps & {
-  triggerLabel: string
-  triggerDisabled?: boolean
-  onClosed?: () => void
-  className?: string
-}
-
-export function QrScannerModal(props: QrScannerModalProps) {
-  const { t } = useI18n()
-  const {
-    triggerLabel,
-    triggerDisabled = false,
-    onClosed,
-    cameraAvailable = true,
-    title: titleProp,
-    stopHint: stopHintProp,
-    className,
-  } = props
-  const title = titleProp ?? t("scanner.defaultTitle")
-  const stopHint = stopHintProp ?? t("scanner.stopHint.modal")
-  const multipartSession = props.multipart?.session
-  const multipartRef = useRef(props.multipart)
-  const contentRef = useRef<HTMLDivElement>(null)
-  const mountedRef = useRef(true)
-  const openRef = useRef(false)
-  const openGenerationRef = useRef(0)
-  const deliveryBusyRef = useRef(false)
-  const automaticCloseRef = useRef(false)
-  const previousClosedKindRef = useRef<TransferState["kind"]>(
-    multipartSession?.state().kind ?? "idle",
-  )
-  const [open, setOpen] = useState(false)
-  const [deliveryBusy, setDeliveryBusy] = useState(false)
-  const [closedNotice, setClosedNotice] = useState<LocalizedText | null>(null)
-  const localizedClosedNotice =
-    closedNotice === null
-      ? null
-      : t(closedNotice.key, closedNotice.values)
-
-  // Take the one-megabyte reader fetch off the acquisition path: warming here means the
-  // binary is normally compiled before the user ever taps start.
-  useEffect(() => {
-    warmQrReader()
-  }, [])
-
-  const beginDelivery = useCallback((): boolean => {
-    if (deliveryBusyRef.current) return false
-    deliveryBusyRef.current = true
-    if (mountedRef.current) setDeliveryBusy(true)
-    return true
-  }, [])
-
-  const endDelivery = useCallback(() => {
-    deliveryBusyRef.current = false
-    if (mountedRef.current) setDeliveryBusy(false)
-  }, [])
-
-  useEffect(() => {
-    multipartRef.current = props.multipart
-  }, [props.multipart])
-
-  useEffect(() => {
-    mountedRef.current = true
-    return () => {
-      mountedRef.current = false
-      openGenerationRef.current += 1
-    }
-  }, [])
-
-  useEffect(() => {
-    previousClosedKindRef.current = multipartSession?.state().kind ?? "idle"
-  }, [multipartSession])
-
-  useEffect(() => {
-    if (open || multipartSession === undefined) return
-    let polling = true
-
-    const inspectClosedSession = async () => {
-      if (
-        !polling ||
-        openRef.current ||
-        deliveryBusyRef.current
-      ) {
-        return
-      }
-      const generation = openGenerationRef.current
-      const canPublish = () =>
-        polling &&
-        mountedRef.current &&
-        !openRef.current &&
-        openGenerationRef.current === generation
-      const previousKind = previousClosedKindRef.current
-      const next = multipartSession.state()
-      previousClosedKindRef.current = next.kind
-
-      if (next.kind === "collecting") {
-        if (canPublish()) {
-          setClosedNotice(
-            localized("scanner.closed.multipartProgress", {
-              received: next.receivedIndexes.size,
-              total: next.frameCount,
-            }),
-          )
-        }
-        return
-      }
-      if (next.kind === "idle") {
-        if (previousKind === "collecting" && canPublish()) {
-          setClosedNotice(localized("scanner.error.expiredDiscarded"))
-        }
-        return
-      }
-      if (next.kind === "error") {
-        if (canPublish()) setClosedNotice(localizedErrorCode(next.code))
-        return
-      }
-      if (!beginDelivery()) return
-      try {
-        if (!multipartSession.claimCompletion()) return
-        const onComplete = multipartRef.current?.onComplete
-        if (onComplete === undefined) return
-        await onComplete({
-          artifactType: next.artifactType,
-          artifactBytes: next.artifactBytes,
-        })
-        if (canPublish()) {
-          setClosedNotice(
-            localized("scanner.closed.integrityImported"),
-          )
-          onClosed?.()
-        }
-      } catch (caught) {
-        if (canPublish()) {
-          setClosedNotice(localizedErrorCode(deliveryError(caught).code))
-        }
-      } finally {
-        endDelivery()
-      }
-    }
-
-    void inspectClosedSession()
-    const timer = window.setInterval(() => {
-      void inspectClosedSession()
-    }, 1_000)
-    return () => {
-      polling = false
-      window.clearInterval(timer)
-    }
-  }, [beginDelivery, endDelivery, multipartSession, onClosed, open])
-
-  const handleOpenChange = (nextOpen: boolean) => {
-    if (nextOpen && deliveryBusyRef.current) return
-    openRef.current = nextOpen
-    if (nextOpen) {
-      openGenerationRef.current += 1
-      automaticCloseRef.current = false
-      previousClosedKindRef.current =
-        multipartSession?.state().kind ?? "idle"
-      setClosedNotice(null)
-    }
-    setOpen(nextOpen)
-  }
-
-  const panelGeneration = openGenerationRef.current
-  const deliverFromPanel = async (
-    generation: number,
-    operation: () => void | Promise<void>,
-    successNotice?: LocalizedText,
-  ) => {
-    if (!beginDelivery()) throw new AppError("INVALID_QR_PAYLOAD")
-    try {
-      await operation()
-      const sameGeneration =
-        mountedRef.current && openGenerationRef.current === generation
-      if (sameGeneration && successNotice !== undefined) {
-        setClosedNotice(successNotice)
-      }
-      if (openRef.current && sameGeneration) {
-        automaticCloseRef.current = true
-        openRef.current = false
-        setOpen(false)
-      }
-    } catch (caught) {
-      if (
-        openRef.current &&
-        mountedRef.current &&
-        openGenerationRef.current === generation
-      ) {
-        throw caught
-      }
-      if (
-        mountedRef.current &&
-        openGenerationRef.current === generation
-      ) {
-        setClosedNotice(localizedErrorCode(deliveryError(caught).code))
-      }
-    } finally {
-      endDelivery()
-    }
-  }
-
-  const panel =
-    props.multipart === undefined ? (
-      <QrScannerPanel
-        singleTargets={props.singleTargets}
-        onSingleScan={(target, payload) =>
-          deliverFromPanel(panelGeneration, () =>
-            props.onSingleScan(target, payload),
-          )
-        }
-        cameraAvailable={cameraAvailable}
-        title={title}
-        autoStart
-        stopHint={stopHint}
-      />
-    ) : (
-      <QrScannerPanel
-        singleTargets={props.singleTargets}
-        onSingleScan={(target, payload) =>
-          deliverFromPanel(panelGeneration, () =>
-            props.onSingleScan(target, payload),
-          )
-        }
-        cameraAvailable={cameraAvailable}
-        title={title}
-        autoStart
-        stopHint={stopHint}
-        multipart={{
-          session: props.multipart.session,
-          onComplete: (completion) =>
-            deliverFromPanel(
-              panelGeneration,
-              () => props.multipart?.onComplete(completion),
-              localized("scanner.closed.integrityImported"),
-            ),
-        }}
-      />
-    )
-
-  return (
-    <div
-      className={cn("space-y-2", className)}
-      aria-busy={deliveryBusy}
-    >
-      <Dialog open={open} onOpenChange={handleOpenChange}>
-        <DialogTrigger asChild>
-          <Button
-            type="button"
-            className="h-11 w-full"
-            disabled={!cameraAvailable || triggerDisabled || deliveryBusy}
-          >
-            <ScanLine aria-hidden="true" />
-            {triggerLabel}
-          </Button>
-        </DialogTrigger>
-        <DialogContent
-          ref={contentRef}
-          tabIndex={-1}
-          className="grid max-h-[95dvh] max-w-lg grid-rows-[auto_minmax(0,1fr)] overflow-hidden p-4"
-          onOpenAutoFocus={(event) => {
-            event.preventDefault()
-            contentRef.current?.focus()
-          }}
-          onCloseAutoFocus={(event) => {
-            if (automaticCloseRef.current) event.preventDefault()
-            automaticCloseRef.current = false
-            if (!openRef.current) onClosed?.()
-          }}
-        >
-          <DialogTitle className="sr-only">{title}</DialogTitle>
-          {/* 4rem prior chrome + ~44px close row + 1rem grid gap */}
-          <div
-            data-qr-scanner-scroll-region
-            className="min-h-0 max-h-[calc(95dvh-4rem)] overflow-y-auto pb-14"
-          >
-            {open && panel}
-          </div>
-        </DialogContent>
-      </Dialog>
-      {!cameraAvailable && (
-        <p className="text-sm text-muted-foreground">
-          {t("scanner.error.cameraUnavailable")}
-        </p>
-      )}
-      {closedNotice && (
-        <p
-          aria-live="polite"
-          className="whitespace-pre-line text-sm text-muted-foreground"
-        >
-          {localizedClosedNotice}
-        </p>
-      )}
-    </div>
   )
 }
