+++ tests/ui/qr-scanner.test.tsx
@@ -10,10 +10,8 @@ import {
 } from "@testing-library/react"
 import userEvent from "@testing-library/user-event"
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest"
-import {
-  QrScannerModal,
-  QrScannerPanel,
-} from "@/components/qr-scanner-panel"
+import { QrScannerModal } from "@/components/qr-scanner-modal"
+import { QrScannerPanel } from "@/components/qr-scanner-panel"
 import {
   Dialog,
   DialogContent,
